Don't place ImplementationController URLs in the AlgorithmController namespace
As agreed on, each controller will get its own version.
In general, the internal URL format is /<controller>/v<version>/...,
this however fails for the previous ImplementationController version,
which placed its endpoints under /algorithm/<algoId>/implementation/...

This change also allows us to omit the algoId in places where it is not necessary,
for example when obtaining a specific implementation.

Also, fix the ImplementationController tests and stop hard-coding URLs in them.