Use a custom LinkBuilder for HATEOAS links
Using the standard `WebMvcLinkBuilder` is not possible because it only inspects
the class- and method-level `@RequestMapping` annotations, ignoring any
custom processing that might happen in a `RequestMappingHandlerMapping` subclass.

This also removes the explicit linkassembler.* instantiations in our tests, as
this becomes more cumbersome when having to instantiate their dependencies as well.