Set the relation names for the algorithm DTO types
Spring HATEOAS uses the link relation names as keys in the `_embedded` resource object.
We need to use the same relation names for all algorithm types, as otherwise we end up
with several lists for the same query, for example:

{
  "classicAlgorithmDtoes": [...],
  "quantumAlgorithmDtoes": [...]
}

This not only breaks sorting but is also incompatible with the generated spec,
which can only consider the declared base type.