Removing hateoas.Links
Links was used in EntityModelConverter class which was in turn used in swagger configuration.  Removed that part of swagger configuration and tested the swagger to find it working fine. So I deleted EntityModelConverter and its test class. Also I find it safe to delete OverrideModelConverter class and its test class also.